Move reports from FakeGithubConnection to github data
This is needed to handle comment pull as direct post request in the
followup. Further the reports are better located in the github data
since this needs to be global data.

Further this will be needed as preparation for the HA scheduler as
well since there will be more than one connection and the test data
needs to be centralized then.

Change-Id: I3fa96d43f0ffe62d50630856b23972221bf2c69a