WIP: Close Socket Connection in Zuul Stream after Stream Ended
We sporadically see "broken pipes" in the kubectl port-forward process
which is opened for log streaming. This apparently can happen if the
client does not properly close its TCP connection piped through the
port-forward, leading to subsequent socket connections to be broken.
(cf. https://github.com/kubernetes/client-go/issues/329)

This is an attempt to work around this by explicitly closing the socket
before terminating the log streamer thread.

Change-Id: Id16cd6a3787697879573323130695f3ee400a4da