PF4: Update builds and buildsets tables
This updates the tables shown on the builds and buildsets pages with new
PF4 components. While the data is loading, a fetching animation is shown
and in case no results could be found the page shows an empty state.

Each table row is clickable and points to the build/buildset result page.
The table uses the same hover effect as the build list on the buildset
result page.

Change-Id: I0f5e0815c53d18e8ae3570dc94594c77fecb90ce