Update patch set 10
Patch Set 10:

Thank you for the details. I don't think `auto reload` should be part of the preferences dialog, as it is only used in the status page, and perhaps it should be removed entirely, I can't think of a use-case where it is useful (beside when debugging or taking a screenshot).

For text wrapping style, it seems to be a per logfile preference, thus it might not be very ergonomic to reach for the user preferences dialog in order to switch the rendering, perhaps that would be better as a toggle directly on the file viewer container?

For ANSI rendering, I can't think of a situation you would want to disable it.

And finally for the number of lines in abridged format, isn't the `more` button enough? And if the default needs to be adjusted, perhaps it should just be adjusted, otherwise the users who do not change their settings won't get the best view. Could a user preferences be set per project or tenant so that everyone benefit from the best viewing experience?


Thus I'm not convinced adding a preferences dialog would improve the Zuul UI, but I'm not a UI expert, so someone else must approve.

Patch-set: 10