Update patch set 4
Patch Set 4:

The length of the input comes with the PF4 toolbar. I have to check if and how I can change this properly.

Autocompletion would be a nice feature indeed. I also had that in mind but as this would require some changes in the backend and the API it's out of scope for this change.

The filter indentation comes with PF4. It should be fine if the other components on the page are updated to PF4 as well. If you take a look at the follow-up change https://review.opendev.org/#/c/746112/1 you can see that the filter toolbar is on the same level as the table contents. So IMHO this should be fine.

Patch-set: 4