Return number of total results from builds/ and buildsets/ API endpoint
This information is necessary to implement a pagination in the zuul web
dashboard. Without this information we don't know how many pages we
should show and when the "next page" button should be disabled.

As this is a backwards incompatible change, one must set the new header
field "Zuul-Api-Version" to "v2" to get the new result format.

When using the Zuul API directly, this information might also come in
handy since otherwise it's unknown how many results one could expect
from the API - especially when applying various filters.

Change-Id: Ib3592b1fc0e56bac165b0dc2825791db49f6ac08