Update patch set 10
Patch Set 10: Code-Review+2

Tristan Cacqueray wrote:

> Thank you for the details. I don't think `auto reload` should be
> part of the preferences dialog, as it is only used in the status
> page, and perhaps it should be removed entirely, I can't think of a
> use-case where it is useful (beside when debugging or taking a
> screenshot).

I think the original use case was people wanting to avoid downloading 5MB files from openstack every 5 seconds.  I agree, I don't think it's that important, but given that use case, I think having a user preference might make sense.  For folks on a slow or expensive connection, they could disable it in preferences and be done.  For everyone else, we get the option off of the status page and they don't have to see it again.

> For text wrapping style, it seems to be a per logfile preference,
> thus it might not be very ergonomic to reach for the user
> preferences dialog in order to switch the rendering, perhaps that
> would be better as a toggle directly on the file viewer container?

I also somewhat agree here, but I think we've seen enough feedback to anticipate that even if we all might switch from one to the other from time to time, that we probably can't really agree on what the default should be.  So maybe the best of both worlds is to have a user pref for it (so you can set a personal default), and also have a temporary switch on the page?  I'd love a hot-key for that too.  :)

> For ANSI rendering, I can't think of a situation you would want to
> disable it.

Me neither.

> And finally for the number of lines in abridged format, isn't the
> `more` button enough? And if the default needs to be adjusted,
> perhaps it should just be adjusted, otherwise the users who do not
> change their settings won't get the best view. Could a user
> preferences be set per project or tenant so that everyone benefit
> from the best viewing experience?

I agree that it seems like "more" should be enough.  I think we can make a small change to the default (ie, if we find that 50 lines is better than 42), but I think it's good right where it is.

This is a case where we should keep in mind one of the maxims we have for Zuul operators: we don't add options just because we can, we try to make it work by default for everyone, and only add options where we really need to because there are substantially different conditions.

> Thus I'm not convinced adding a preferences dialog would improve the
> Zuul UI, but I'm not a UI expert, so someone else must approve.

Sorin has 4 suggestions, I agree with 2 of them (this one included), and Tobias has a +2.  I think we're leaning toward consensus that the user prefs dialog is good, and we can add the auto-reload toggle to it.  I'm going to hold off on +W for a little longer in case anyone else wants to weigh in.  I'd also like Clark to re-review, and Felix has some unresolved comments.

Patch-set: 10
Label: Code-Review=+2