Update patch set 10
Patch Set 10:

I will address all notes, including the removal of custom styles in follow up, mainly because it would make the change much easier to review and also unblock other changes which are waiting for this one.

IMHO, we cannot make this one perfect from 1st try but we will have opportunity to improve it before the release.

Patch-set: 10