WIP web: Implement basic fuzzy search for job_name on builds page
When searching for builds in Zuul it's sometimes inconvinient to type in
the exact job name that you want to search. In most cases only a part of
the name is sufficient. In case you want to search for multiple jobs
that all have the same prefix, you also don't want to specify each exact
job_name as query parameter.

This improves the search by utilizing SQL's regexp operator when
filtering for the provided job_name.

Change-Id: I0ec42f9f595411ac1e5ef6cc6b8222d6164c1e47