WIP: configloader: do not run getProjectBranches when load_classes empty
For the moment this commit is just a test to explore possibility
to defer the getProjectBranches after the tenant loading process
for projects with no config classes (include: []).

Indeed in case of tenant with thousands of repos with include: []
we would like to avoid long startup phase where thousand
of API calls are performed to get projects branches.

Change-Id: I9f056a99292010a1b662dbc247e6d7b4875cdefe