Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

I'm not sure if this is sufficient to be really useful for the end user. The problem is that this is still part of the zuul package and thus comes with all dependencies that a fully featured zuul needs. For example on MacOS the zuul package is not even installable.

I wonder if we need to split this off into its own project?

Patch-set: 10
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1