Update patch set 6
Patch Set 6:

> (1 comment)

Hi Ian,

thats because by this change we do not longer rely on zuul pull request representation to get a reviewers info, but the corresponding user object (cf. FakeUser in fakegithub).
But you're right, this is a confusing leftover. I changed that to '@example.com' for the sake of consistency.

Patch-set: 6
Label: Verified=0