Add always-files option to jobs
This adds an "always-files" options to jobs.  Unlike the regular
"files" matcher, this is inherited by children, and and extra values
are appended through the inheritance chain.

An example of where this would be useful is the various service jobs
in OpenDev system-config. These jobs all inherit from a common base
job, but limit themselves with file matchers to only run on specific
changes to the roles they use.  The base job installs Ansible and
takes the default version from a defaults file in the included
install-ansible role.  If we update the Ansible version there, we want
all child jobs to run and validate they work with the new Ansible.

Currently each individual child job needs to remember to add a file
matcher for this role/defaults file to trigger on such an update.  The
worst case is that child jobs forget the match, aren't tested with the
new version and break when you commit it.  On a more theoretical
level, having to understand what roles are used in the parent job to
construct matchers breaks the abstraction; child jobs don't (and
shouldn't) really care how Ansible is installed there, just as long as
it is installed.

I think this makes the case for file-matchers that are passed through
inheritance.  The idea is in essence simple; entries from
"always-files" act as though they are appended automatically to each
jobs "files" matcher list.

Change-Id: Iec596b9f967d5a6080874e9e438db3ce8bd76e5c