Update patch set 5
Patch Set 5: Code-Review-1

That needs more work because in its current form it downgrades the UX by creating more wrapping.

https://sbarnea.com/ss/Screen-Shot-2020-08-24-12-01-11.01.png


Repeating the same icon on each cell of a table adds zero value and increases the horizontal space needed, causing extra wrapping.

Almost none of the cells should have icons. The only case were an icon could make sense to have would be if it can have multiple values or missing it.


Based on what I see the only places where it makes sense is the first column, the build because the icon can be a "v", "!" or "x".

Patch-set: 5
Label: Code-Review=-1