Update patch set 4
Patch Set 4:

zbr: First of all, this change only adds the filter toolbar. It doesn't change the table in any way. That's done in the follow up change https://review.opendev.org/#/c/746112/5. So please, don't complain about anything here that's not part of this change.

Second: What exactly do you mean by "That needs more work because in its current form it downgrades the UX by creating more wrapping."? I don't think that this change "downgrades" the UX in any way. Regarding the "additional" padding, I've always responded that this is done by PF4 directly and IMHO fits the overall layout with the new table that's part of the follow up change.

Regarding the icons: I will address this in the change where it belongs.

Patch-set: 4