Update patch set 5
Patch Set 5:

ianw: I see, for fields where the values often repeat (pipeline, project, branch) the combination wit the icon indeed amplifies this effect. Another idea would be to add the icons only to the table headers, but AFAIK that's not possible as PF4 tables only accept simple string values as column headers. I will check this again and see if there is another way to achieve this.

Patch-set: 5