Update patch set 6
Patch Set 6:

I am using a 5K retina display with a symmetrical split in two. This makes the browser width ~1260px.

I think that on 4K display this would be even worse, as it would likely produce a ~1000px "columns".

I do not expect zero wrapping, but clearly not unavoidable one.

Patch-set: 6