Update patch set 4
Patch Set 4: Code-Review+1

There's a break at 1200 pixels wide that changes this into a folded widget that needs to be expanded.  Considering that the width of the widget itself is only about 420 pixels wide, that seems excessive and unintuitive.  I noticed this because I just happened to load this up in a window which was not full-screen and I didn't understand the change (I just saw that the filter bar had been turned into a filter icon, and I didn't understand that I should click the icon to expand the widget; I might expect that on mobile, but not on a relatively wide window on desktop).

I understand that the break is system-wide -- I see the navbar changes at the same breakpoint, and therefore, adjusting this might be out of scope for this change.  But before we merge it, would you mind taking a look into that to see if there's something that could be done?  (Perhaps something about setting the width to 100% causes this behavior?)

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1