Update patch set 6
Patch Set 6: Code-Review+2

Speaking of wrapping, we might want to take a look at whether we can reduce the wrapping which is evident in the 'openstack' and 'pyca' tenants on opendev.  But the latest revision of this change doesn't seem to make it worse, so lgtm.

Patch-set: 6
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2