Update patch set 6
Patch Set 6: Code-Review-1

> If you hover over the pipeline column header you will endup seeing react code as tooltip. That is clearly something new.

I can not replicate that ...

The builds page looks great to me, with the icons at the top. however, the buildsets page now has the icons next to each line.  I think we should update this to be consistent and just put the icons in the headers.

Consider me +2 modulo that

Patch-set: 6
Label: Code-Review=-1