Update patch set 2
Patch Set 2: Code-Review-1

This change is weird when using a screen reader in that we set the focus to be in the middle of the page, and it's not obvious that a user should navigate backwards to access the navigation menu.  The current behavior is more natural, in that the focus is on the logo, and tab-navigation goes through the various links in the navbar.

We might be able to mitigate this by improving our navigation (via correct aria roles, or headers, or similar).  I'm not sure what the best approach is, but I think if we want to do something unusual like this, we need to compensate.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1