Update patch set 2
Patch Set 2: Code-Review+2

Considering the problems with tab navigation we found in Felix's changes, perhaps this is the best way to go.  This seems like a much more minor regression (and indeed, there are some who prefer the navbar scroll out of view, so for those folks, it's a double win).

If we find a way to restore the intended navbar behavior without disrupting tab navigation or introducing other problems, I'd be in favor of that.  But for the moment, this seems like either the least-bad or even best option depending on perspective.

Patch-set: 2
Label: Code-Review=+2