Update patch set 9
Patch Set 9: Code-Review-1

When showing long log lines, the horizontal scrollbar should always be visible in the browser.  The user should not have to scroll down to the bottom of the container to find the horizontal scroll bar in order to scroll right.  With this patch, that is the case for both the "results" tab and the "console" tab.  That may not be the fault of this patch -- I think there's a weird interaction with the PF4 tabs we currently have.  I think we're about to revert the pf4 tabs change, so once we get that straigtened out, this patch may be in better shape.

When opening a log from the "logs" tab, if wrapping is enabled, then the line number column is only one character wide, so the actual log lines end up having a lot of space between them.

Patch-set: 9
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1