Exercise github auth handling in tests
So far almost all of the github auth handling has been overridden in
the fake test classes. We can increase the test coverage and gain more
confidence when doing changes related to this by not replacing
getGithubClient. Instead call the super method and enhance the
resulting client with the relevant test data.

Change-Id: Icaa9cfb9f28f8faec2dc86f7e6393f5b23765cac