[dnm] test squashed web revert & modal fix
Revert "web: restore scrollbars and scroll behaviour"

This accidentally broke the layout of the build result page and the tabs only reachable by scrolling horizontally in case a build log contained very long log lines.

This reverts commit 8d0f440e1bc85b55a2d780769d8ee1df4ae6ce7e.

Change-Id: Ifcdf5c737ef6f2e178c9a4531ead81e8c0fe6505

Revert "Revert PF4 build page"

The original change was reverted to fix a scrolling bug that was introduced by a different change. Using this commit in combination with [1] should restore the old behaviour.

Applying [2] on top of them should finally get rid of the scrolling issues.

[1]: https://review.opendev.org/#/c/750361/
[2]: https://review.opendev.org/#/c/750322/

This reverts commit b4b5b9fd58dd5d5aeb9d32331e4882d7ef5faf06.

Change-Id: Icd498314762a8edca751413b7ee07b9b72317c5b

Use Modal to show config errors and fix scrolling

Since we introduced PF4 together with the Drawer component to show the
config errors, we are not able to scroll the page by using the keyboard.
So far, no workaround really did the trick and every one came with
another drawback in different areas.

So we decided to ditch the Drawer component and use a Modal as
replacement. This should still serve the use case as we have "something
popping up when clicking on the bell icon", but this something doesn't
break the layout and/or the scrolling behaviour.

Additionally this change removes the "height: 100%" CSS rule on the root
element. In combination with the removed drawer this should restore the
scrolling behaviour via keyboard without introducing any further
drawbacks like wrong tabbing or focus behaviour.

For additional background information on this topic, please refer to the
changes [1], [2] and [3]. There is also a PF4 issue on Github available
[4].

[1] https://review.opendev.org/#/c/743239/
[2] https://review.opendev.org/#/c/743917/
[3] https://review.opendev.org/#/c/742759/
[4] https://github.com/patternfly/patternfly-react/issues/4624

Change-Id: I45df9cabc96d2d7fb4ddebdf1c72d0dd8a3adfd2