Update patch set 2
Patch Set 2: Code-Review-1

I like the modal as a solution.  I think we need to set the text to "pre-wrap" in order for the text to look right.

If you load up a failing build with long lines, you'll note the following is true for the task output in both the summary and console tabs: the horizontal scroll bar is attached to the div, not the viewport for the page.  This means that the user needs to scroll down to the bottom of the div to scroll right, and then scroll back up to the line they wanted to see.  I think this is unworkable and we need to either maintain the current behavior where the horizontal scroll bar is attached to the viewport (ie, it is always visible at the bottom of the screen), or we should wrap text in both of these cases.  If we chose the latter, I do think we should still have viewport-level horizontal scrolling for the logs tab (so while switching to wrapped text may reduce the problem space and make things simpler, I'm not sure it eliminates the problem completely).

Finally, the logs tab in this change has text that is wider than the screen but no horizontal scroll bar at all.  That seems strange.

Patch-set: 2
Label: Code-Review=-1