Update patch set 2
Patch Set 2:

Just to add some comments

https://i.imgur.com/MlK1s8h.png

shows the horizontal scrolling issue I believe.  The problem is if it's a long line at the top, it has scrolled off the screen when you've moved down the bottom to find the scroll-bar.

Looking at other contemporary interfaces, they all seem to wrap.  For example

https://github.com/pyca/cryptography/pull/5428/checks (you need to be logged in, but then you can pull down the arrows)

https://gitlab.com/wireshark/wireshark/-/jobs/727727167 (shrink window, but there's some long lines there)

To me, that's enough to suggest that wrapping is OK in this case.

I also agree that there's no way to view the long lines on the log viewing page.  Even when you select the text, it doesn't scroll the viewport over.  I am guessing this has to do with every line being a table row and something to do with table views?  Currently, it looks like the whole page gets wider.

Patch-set: 2
Label: Workflow=0