Update patch set 2
Patch Set 2:

No objections on the pre-wrap from my side, but I don't get the scroll problem here:

Providing the scroll bar for the whole viewport is exactly the problem why the page was wider than the browser could show in a window - otherwise you won't get the viewport scrollbars. By removing the overflow: visible attributes, we can keep the scrollbars inside those containers which need it - which is the most common solution IMHO (e.g. showing code snippets on websites that have longer lines do the same).

So, if we go for the pre-wrap here I'm fine with that. I'm still wondering how that looked before (with PF3). If the text wasn't wrapped there either it would have also provided scrollbars inside the tab content - which didn't seem to bother anybody at that time. I will check that out.

Patch-set: 2