Fix memleak on zk session loss
When the scheduler looses its zk session it resubmits all lost node
requests as new ones. However it didn't stop the watch for the old one
which keeps being registered in Kazoo. The watch contains the
NodeRequest object since it's bound to the callback. Thus by leaking
the Watch we also leak the NodeRequest, the attached BuildSet,
QueueItem and finally the Tenant and Layout as well.

This can be fixed by stopping the watch in this case.

Change-Id: I3b05ec92816ab5eb06ad40dfad85ddfebfbf2cc4