UI: Wrap lines on Logfile page
PF4 broke the horizontal scrolling behaviour of the pre/table
combination used on this page. But, as we don't like horizontal
scrolling anyway, go for a pre-wrap solution on the log messages. Due to
the line number shown in front of the message, this looks quite nice and
still provides the context of each line.

Change-Id: I038942d235fd13cae90ed82429558f441f1dcbb4