web: PF4 minor rework of log viewer page
I think the log-viewer page could do with some PF4-ness.

This moves the title section into a card from a bootstrap panel, and
puts the filename you're looking at in the prominent location and the
build-id in a less prominent location.

The filters are moved from jammed-together links into a button
toggle-group, which is a mutually-exclusive set perfect for this
interface.  This component requires the latest release of PF4, which
is why the packages have been updated.

Wrap the output in a PgaeSection that spaces things out a bit more
nicely.

Change-Id: Ibcbc2bd9497f1d8b75acd9e4979a289173d014b2