web: PF4 minor rework of log viewer page
I think the log-viewer page could do with some PF4-ness.

This moves the title section into a card from a bootstrap panel, and
puts the filename you're looking at in the prominent location and the
build-id in a less prominent location.

The filters are moved from jammed-together links into a button
toggle-group, which is a mutually-exclusive set perfect for this
interface.  This component requires the latest release of PF4, which
is why the packages have been updated.

The filtering is slightly reworked so that clicking on a button sets
the "severity" state variable, instead of the Link model setting props
before.

Because the filters are, I think, a bit of an OpenStacky thing
(devstack type logs) and generally not constantly used, I put them in
a drop-down box so they don't take up space when not relevant.

Wrap the output in a PgaeSection that spaces things out a bit more
nicely.

Change-Id: Ibcbc2bd9497f1d8b75acd9e4979a289173d014b2