Clear traceback before attaching exception to event
When we hit an exception in a management event we attach it to the
event so the sender of the event can get the stack trace back. However
the traceback also contains the local variables of all frames part of
the traceback. Those also can contain references to Tenant objects
which will be blocked for the gc until the event is (hopefully)
collected as well. Since we're only interested in the stack traces and
don't need the local variables clear the frames before attaching the
traceback to the event.

Change-Id: I57854ac4b9d1405d2b7b86f27149f87af968e77a