Remove source_event from Change objects
In the past source_event has been used to improve merge messages. This
is being attached to changes in getChange. Since there can be any
event on a change this has been wrong and is now unused. Thus remove
this variable. Further it currently even creates a problematic memleak
when doing dequeue commands without the change being in any queue. In
this case the exception trace including all local variables of the
frames (e.g. Tenant) are attached to the event and thus to the
cached change which will never be cleared unless there is another
event incoming for this change).

Change-Id: I23e59e66f485bf46493d7a5bdf7bc29966e56993