Ignore 500 errors when requesting pr files
Github responds with a 500 error if the diff is too large to make
sense. This is currently handled by the retry handler with backoff and
can significantly delay event processing. However Github will never
respond with something different so this needs to be excluded from the
retry handler.

Further it has been revealed that our error handling code incorrectly
results in an empty changed files list. However in this case we must
set the changed files to None so zuul takes care of this itself.

Change-Id: Ie825a8801032d5a1ec66afb1244aa3b571fd1d39