Correct visibility check in web JS
A recent change to introduce user preferences in the client-side
web app had an error when detecting the case where a user switched
from having auto-reload of the status page disabled to enabled.
It only checked that the autoreload setting was enabled and there
was no timer currently set; if that was true, then it would force
a refresh.  This check was performed every time the component was
updated, so each time a status update happened, the component would
naturally be updated as part of the process, run this check, see
that autoreload was enabled, and then force an update regardless
of the visibility setting.

This change ensures the forced update only happens in the specific
case that the user has changed the autoreload setting, thereby
breaking the loop described above.

Change-Id: Ie7ef175536bb20033cede5e93748cda08c3d918d