Update patch set 1
Patch Set 1:

> 'approved' being true when nobody approved a merge-request seems to me to be an unexpected/surprising 
> behavior, even if this merge-request does not require any approval.

Just to clarify, your are not seeing 'approved' at True when approvals are requested but not satisfied right ?

Your are seeing it as True when no approval is requested or when all approvals are satisfied ?

If yes, then that's seems correct to me, even if that's true it's odd to tell a change is approved if no approval is requested :) 

> Using 'approved' would require an update of the documentation related to the Zuul gitlab driver. What do you think ?

If we just choose to use 'approved' field, then this internal change, seems to be technical detail. I don't see what to add/change in the documentation. But let me know, what would you like to add/change ?

Patch-set: 1