Update patch set 4
Patch Set 4:

First impression: IMHO we should show the log itself inside the tab (similar to the file tree). This way we can still see the general build attributes and you directly know to which build the log belongs. I know that this wasn't like that before, but I always missed that when viewing a log on this page. Additionally we can keep the page layout when switching between the different tab contents.

I totally like the design of the new filter bar.

Maybe it would make sense to show some breadcrumbs above (or below?) the filter bar showing the path to the current logfile. As we could have arbitrarily nested log files, this would help to keep track on which file in the tree you currently viewing.

Patch-set: 4
Label: Verified=0