Update patch set 5
Patch Set 5:

(2 comments)

> Sorry I don't 100% follow; you're saying that when we click on the link for a log in the existing "logs" tab, it does not open in a fresh page, but replaces the contents of the "logs" tab with the contents of the log file?

Yes, I would "replace" the tab content with the log file. Since the LogFile is currently a dedicated page, we might have to think how to integrate it as a container in the Build page and only show it once a logfile is selected. AFAIK this decision is currently done by the router. So we could do something like this https://review.opendev.org/#/c/750365/3/web/src/routes.js also for the logs/:filename route, set the activeTab to "logs" and specify the filename as additional property for the build page. Then in the build page we could show the single log file content in the logs tab if the filename attribute is set.

Patch-set: 5
Label: Verified=0