Save superfluous api requests in check run reporting
When reporting the check run result zuul is currently doing four
github api requests:
1. Get repository
2. Get head commit of pr
3. Get check runs of head commit
4. (optional) Create initial version of check run in case start
   reporting is disabled
5. Update check run that has been created in start reporting

The first three requests are basically unneeded if we craft a direct
PATCH request and remember the check run id that has been created in
the start reporting.

This is needed since those run in the critical path of the event
processing and can cause large event processing delays in a very busy
zuul system.

Change-Id: I55df1cc28279bb6923e51686dde8809421486c6a