Make revising node requests asynchronous
When having large queues in the system and a change at the front
finishes all node requests of later changes are updated due to changed
relative priorities if enabled. This can have a significant
performance impact on event processing since this is done
synchronously within the main loop. The update of each node request
consists of three network round trips to zookeeper (lock, update,
unlock). Since we ignore errors in this case we can just offload this
to a ThreadPoolExecutor which does this in a fire and forget manner.

Change-Id: I00d56ed43151a9201953693d35ac685686249859