Update patch set 1
Patch Set 1: Workflow-1

This change is just a POC, but is not actively working because we are blocked by felix changes on pf4. I will resume working on it once these are merged.

This is in order to avoid conflicts.


PS. I am happy to see people looking a it, the goal was to get some initial feedback, hopefully being able to avoid a lets make it configurable approach.

Regarding durations, I would personally even use "00:00" duration.

We are currently using moment.js which is kinda of end-of life, so we may use the opportunity to switch to something else.


Keep in mind that we are no dropping precision, we still display the exact values as tooltips, that is important.

Patch-set: 1
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1