Update patch set 8
Patch Set 8:

> Yes, I would "replace" the tab content with the log file

That sounds interesting; I mean it would be kind of nice if it opened in separate tabs and you could flick between them.  But I wonder if a few megabyte logs might start to make things slow to a crawl...

I did try to somehow route in the current path to the this view to give more breadcrumbs on the filename, but so far my react skills weren't up to the challenge.  maybe if it's something quick for you to hack we could follow on with a better way to at least get back to the build-info page when looking at a log, as a first step

Patch-set: 8
Label: Verified=0