web: consistent font sizes on console page and PF4 for task summary
PF4 has highlighted to me we have a range of different font sizes in
the "results" and "console" tab.  I think this stands out a lot more
now because you have the header section with consistent font sizes,
then we start shrinking various things below that.

This removes the Panel font-sizes from CSS so we just keep the
defaults, which to my eye look good.

However, Panel's aren't in PF4, so I've reword the results page to use
Cards.

The host/task summary card is the same generally, but has some bubble
labels and I think the colors make more sense now (green for OK, for
example).  Icons remain the same, but instead of hiding what the value
is in a tool-tip it is included in the label.  I also moved it to be
first always, as it is supposed to be an overview.

If there is are task failures, they are in cards too.  Again this is
basically the same; but I put the host it fails on in a bubble that
matches the "failed" bubble from the summary for consistency, and made
the title a bit clearer explaining what follows is the failed task
output.

I've also renamed the tab title "Task Summary", which to me conveys
more exact info than just "Results" which could be all sorts of
things.

Change-Id: I7b56efbfc9c25872e399853045e9d08f38a0a3bb