web: consistent font sizes on console page and PF4 for task summary
PF4 has highlighted to me we have a range of different font sizes in
the "results" and "console" tab.  I think this stands out a lot more
now because you have the header section with consistent font sizes,
then we start shrinking various things below that.

This removes the font-sizes from CSS so we just keep the defaults,
which to my eye look good.

I've reworked the results tab to use the PF4 datalist.  I put it in a
card and reduced the width with a grid so the results aren't so far
from the hostname.  The info is moved into bubble labels, and I think
the colors make more sense now (green for OK, for example).  Icons
remain the same, but instead of hiding what the value is in a tool-tip
it is included in the label.  I've also called it the "Task Summary",
which to me conveys more info than just "Results" which could be all
sorts of things.

Change-Id: I7b56efbfc9c25872e399853045e9d08f38a0a3bb