Optimize GitHub requests on PR merge
Using github3py for merging a pull request there is always one extra
request to get the pull request before the merge can be done due to the
github3py data model.

Since merging is done synchronously in in the scheduler main loop, it
makes sense to save the extra API request by building and executing the
merge call directly.

Change-Id: I48dbb5fc1b7e7dec4cb09c779b9661d82935d9df